chore: removed .env file, database files, schema files because the website will have the blogposts stored locally instead of on a database
The decision to do this was mainly because, I could not find a way to have Svelte Components and codeblocks render when fetched from a database. They would just render as strings. A work around is possible but after speaking to some experienced develoeprs, I was informed that a site like mine doesn't really need a database